requirements.txt
audiomentations==0.20.0
librosa==0.8.0
matplotlib==3.3.3
numpy==1.19.5
pandas==1.2.0
parselmouth==1.1.1
pymongo==4.0.1
python-dotenv==0.19.2
scikit_learn==1.0.2
scipy==1.6.0
